Move Colors and Typography to Styles
