+++ react/components/EditorContainer/StoreEditor/Styles/components/Colors.tsx
